+++ src/Requests/Orders/GetOrderInvoiceRequest.php
@@ -1,0 +1,27 @@
+<?php namespace Ordercloud\Requests\Orders;
+
+use Ordercloud\Support\CommandBus\Command;
+
+class GetOrderInvoiceRequest implements Command
+{
+    /**
+     * @var int
+     */
+    private $orderId;
+
+    /**
+     * @param int $orderId
+     */
+    public function __construct($orderId)
+    {
+        $this->orderId = $orderId;
+    }
+
+    /**
+     * @return int
+     */
+    public function getOrderId()
+    {
+        return $this->orderId;
+    }
+}